just added the python file to downloads

+++ cleaner.py
@@ -1,6 +1,6 @@
 import os
 
-downloads = "???" #Add the filepath to your downloads folder here
+downloads = "C:/Users/ConorBailey/Downloads/" #Add the filepath to your downloads folder here
 
 if os.path.exists(downloads) and os.path.isdir(downloads):
     files = os.listdir(downloads)
@@ -8,7 +8,7 @@ if os.path.exists(downloads) and os.path.isdir(downloads):
     for file in files:
         file_path = os.path.join(downloads, file)
         try:
-            if os.path.isfile(file_path) and file.lower() != "desktop.ini": # desktop.ini is too important to delete
+            if os.path.isfile(file_path) and file.lower() != "desktop.ini" and file.lower() != "cleaner.py.lnk": # desktop.ini is too important to delete
                 os.remove(file_path)
                 print(f"Deleted: {file_path}")
         except Exception as e:
